Add bootstrap library for navbar
