+++ client/package.json
@@ -10,6 +10,7 @@
     "antd": "^4.10.0",
     "antd-img-crop": "^3.13.2",
     "axios": "^0.21.1",
+    "bootstrap": "^4.5.3",
     "is-empty": "^1.2.0",
     "jwt-decode": "^3.1.2",
     "prop-types": "^15.7.2",
